- True fix https://github.com/HellFirePvP/ModularMachinery/issues/176. - Improvement of TOP energy display, realized https://github.com/HellFirePvP/ModularMachinery/issues/190.

+++ src/main/java/hellfirepvp/modularmachinery/common/util/MiscUtils.java
@@ -123,7 +123,7 @@ public class MiscUtils {
         }
     }
 
-    public static String formatDecimal(float value) {
+    public static String formatDecimal(double value) {
         return DECIMAL_FORMAT.format(value);
     }
